add code stuff

+++ upgrade-control-plane.sh
@@ -14,6 +14,10 @@ sh upgrade-kubectl-kubeadm.sh
 for node in $(kubectl get nodes -o name)
 do
 	node=${node##*/}
+	if(( ${node}=="vm-master" ))
+	then
+		continue
+	fi
 	kubectl cordon ${node}
 	kubectl drain ${node} --ignore-daemonsets
 	ssh azureuser@${node} 'bash -s' < upgrade-kubectl-kubeadm.sh
